requirements.txt
boto3>=1.34.0,<2
python-dotenv>=1.0.0,<2
tabulate>=0.9.0,<1
